Controllers restructured
