+++ cosmetics-web/app/controllers/landing_page_controller.rb
@@ -1,26 +1,28 @@
-class LandingPageController < ApplicationController
-  skip_before_action :authenticate_user!
-  skip_before_action :ensure_secondary_authentication
-  skip_before_action :require_secondary_authentication
-
-  before_action :redirect_poison_centre_users
-  before_action :set_responsible_person
-
-  def index
-    if search_domain?
-      render "search_landing_page"
-    else
-      render "submit_landing_page"
-    end
-  end
-
-private
-
-  def redirect_poison_centre_users
-    redirect_to poison_centre_notifications_path if poison_centre_or_msa_user?
-  end
-
-  def set_responsible_person
-    @responsible_person = current_user.responsible_persons.first if user_signed_in?
-  end
-end
+# class LandingPageController < ApplicationController
+#   skip_before_action :authenticate_user!
+#   skip_before_action :ensure_secondary_authentication
+#   skip_before_action :require_secondary_authentication
+#
+#   before_action :redirect_poison_centre_users
+#   before_action :set_responsible_person
+#
+#   def index
+#     return redirect_to dashboard_path if user_signed_in?
+#
+#     if search_domain?
+#       render "search_landing_page"
+#     else
+#       render "submit_landing_page"
+#     end
+#   end
+#
+# private
+#
+#   def redirect_poison_centre_users
+#     redirect_to poison_centre_notifications_path if poison_centre_or_msa_user?
+#   end
+#
+#   def set_responsible_person
+#     @responsible_person = current_user.responsible_persons.first if user_signed_in?
+#   end
+# end
