+++ cosmetics-web/app/controllers/publicly_accessible_controller.rb
@@ -1,5 +1,3 @@
 class PubliclyAccessibleController < ApplicationController
   skip_before_action :authenticate_user!
-  skip_before_action :has_accepted_declaration
-  skip_before_action :create_or_join_responsible_person
 end
