+++ cosmetics-web/app/controllers/responsible_persons/notifications_controller.rb
@@ -1,6 +1,6 @@
 require "will_paginate/array"
 
-class ResponsiblePersons::NotificationsController < ApplicationController
+class ResponsiblePersons::NotificationsController < SubmitApplicationController
   before_action :set_responsible_person
   before_action :set_notification, only: %i[show edit confirm]
