+++ cosmetics-web/app/controllers/search/dashboard_controller.rb
@@ -1,0 +1,5 @@
+class Search::DashboardController < SearchApplicationController
+  def show
+    redirect_to(poison_centre_notifications_path)
+  end
+end
