+++ cosmetics-web/app/controllers/users/password_changed_controller.rb
@@ -1,7 +1,5 @@
 module Users
   class PasswordChangedController < ApplicationController
-    skip_before_action :create_or_join_responsible_person
-
     def show; end
   end
 end
