+++ cosmetics-web/app/controllers/users/sessions_controller.rb
@@ -2,10 +2,7 @@ module Users
   class SessionsController < Devise::SessionsController
     skip_before_action :authorize_user!
     skip_before_action :set_raven_context
-    skip_before_action :has_accepted_declaration
-    skip_before_action :create_or_join_responsible_person
     skip_before_action :require_secondary_authentication
-    skip_before_action :try_to_finish_account_setup
 
     def create
       set_resource_as_new_user_from_params
