+++ cosmetics-web/app/views/search/landing_page/index.html.erb
@@ -1,0 +1,60 @@
+<% content_for :page_title, "Landing Page" %>
+<div class="app-masthead">
+  <div class="govuk-width-container">
+    <%= render "shared/phase_banner" %>
+
+    <div class="govuk-grid-row">
+      <div class="govuk-grid-column-two-thirds">
+        <h1 class="govuk-heading-xl govuk-!-margin-top-4 govuk-!-margin-bottom-6">
+          Search for cosmetic products
+        </h1>
+        <p class="masthead-account-links govuk-!-padding-top-2 govuk-!-margin-bottom-5">
+          <%= link_to "sign in", new_search_user_session_path, class: "govuk-link govuk-link--no-visited-state" %>
+        </p>
+      </div>
+    </div>
+  </div>
+</div>
+
+<div class="govuk-width-container">
+  <main id="main-content" class="govuk-main-wrapper app-main-class" role="main">
+    <div class="govuk-grid-row">
+      <div class="govuk-grid-column-two-thirds">
+        <p class="govuk-body-lead">Search for cosmetic products available in the UK and notified with the Office for Product Safety and Standards (OPSS).</p>
+
+        <h2 class="govuk-heading-m">How to access the service</h2>
+
+        <p class="govuk-body govuk-!-margin-bottom-2">If you’re using this service for the first time, send an email to <a href="mailto:opss.enquiries@beis.gov.uk">opss.enquiries@beis.gov.uk</a> to ask for an account.</p>
+
+        <h3 class="govuk-heading-s">Trading Standards officers</h3>
+
+        <h2 class="govuk-heading-m">What you can search for</h2>
+
+        <p class="govuk-body govuk-!-margin-bottom-2">You can search for cosmetic products by:</p>
+
+        <ul class="govuk-list govuk-list--bullet">
+          <li>product name</li>
+          <li>product category</li>
+          <li>responsible person</li>
+        </ul>
+
+        <div class="govuk-inset-text">The ‘responsible person’ (RP) is the company (or individual) responsible for ensuring the regulations are followed. Every cosmetic product placed on the UK market must have an RP.</div>
+
+        <h3 class="govuk-heading-s">National Poison Information Service (NPIS) staff</h3>
+
+        <p class="govuk-body">You can find out the ingredients in a cosmetics product.</p>
+
+        <h3 class="govuk-heading-s">Trading Standards officers</h3>
+
+        <p class="govuk-body">You can check if the cosmetic product has been notified with OPSS.</p>
+
+        <h2 class="govuk-heading-m">What you can't search for</h2>
+
+        <p class="govuk-body">You can't search for cosmetic products taken off the UK market before 1 February 2020 but still used in the UK.</p>
+
+        <div class="govuk-inset-text">Some ingredient information and packaging images may be missing for cosmetic products notified with the EU before 1 February 2020.</div>
+
+      </div>
+    </div>
+  </main>
+</div>
