+++ cosmetics-web/app/views/users/password_changed/show.html.erb
@@ -5,6 +5,6 @@
   <div class="govuk-grid-column-two-thirds-from-desktop">
     <h1 class="govuk-heading-l"><%= title %></h1>
 
-    <p class="govuk-body"><%= govukButton href: root_path, text: "Continue" %></p>
+    <p class="govuk-body"><%= govukButton href: dashboard_path, text: "Continue" %></p>
   </div>
 </div>
