+++ cosmetics-web/spec/controllers/search/dashboard_controller_spec.rb
@@ -1,25 +1,6 @@
 require "rails_helper"
 
-RSpec.describe DashboardController, type: :controller do
-  describe "When signed in as a Responsible Person user" do
-    let(:responsible_person) { create(:responsible_person, :with_a_contact_person) }
-
-    before do
-      sign_in_as_member_of_responsible_person(responsible_person)
-    end
-
-    after do
-      sign_out(:submit_user)
-    end
-
-    describe "GET #show" do
-      it "redirects to cosmetic products page" do
-        get :show
-        expect(response).to redirect_to(responsible_person_notifications_path(responsible_person))
-      end
-    end
-  end
-
+RSpec.describe Search::DashboardController, type: :controller do
   describe "When signed in as a Poison Centre user" do
     before do
       sign_in_as_poison_centre_user
